encore des corrections

+++ vm/src/cw_ins_get_data.c
@@ -45,12 +45,12 @@ unsigned int		cw_get_data_dir(t_proc *proc, t_champion *champ,
 unsigned int		cw_get_data_ind(t_proc *proc, t_champion *champ,
 					unsigned int sze)
 {
-	unsigned int	ret;
-	unsigned int	ind;
+	short			ret;
+	short			ind;
 
-	ind = (proc->mem[sze % MEM_SIZE] << 8 | proc->mem[(sze + 1) % MEM_SIZE])
+	ind = (short)(proc->mem[sze % MEM_SIZE] << 8 | proc->mem[(sze + 1) % MEM_SIZE])
 	% IDX_MOD;
-	ret = proc->mem[(champ->pc + ind) % MEM_SIZE] << 24
+	ret = (short)proc->mem[(champ->pc + ind) % MEM_SIZE] << 24
 		| proc->mem[(champ->pc + ind + 1) % MEM_SIZE] << 16
 		| proc->mem[(champ->pc + ind + 2) % MEM_SIZE] << 8
 		| proc->mem[(champ->pc + ind + 3) % MEM_SIZE];
@@ -60,11 +60,11 @@ unsigned int		cw_get_data_ind(t_proc *proc, t_champion *champ,
 unsigned int		cw_get_data_ind_l(t_proc *proc, t_champion *champ,
 					unsigned int sze)
 {
-	unsigned int	ret;
-	unsigned int	ind;
+	short			ret;
+	short			ind;
 
-	ind = (proc->mem[sze % MEM_SIZE] << 8 | proc->mem[(sze + 1) % MEM_SIZE]);
-	ret = proc->mem[(champ->pc + ind) % MEM_SIZE] << 24
+	ind = (short)(proc->mem[sze % MEM_SIZE] << 8 | proc->mem[(sze + 1) % MEM_SIZE]);
+	ret = (short)proc->mem[(champ->pc + ind) % MEM_SIZE] << 24
 		| proc->mem[(champ->pc + ind + 1) % MEM_SIZE] << 16
 		| proc->mem[(champ->pc + ind + 2) % MEM_SIZE] << 8
 		| proc->mem[(champ->pc + ind + 3) % MEM_SIZE];
