+++ vm/src/cw_ins_ldi.c
@@ -54,7 +54,7 @@ int			cw_ins_ldi_firstparamhandler(t_proc *proc, t_champion *tmp,
 
 int			cw_ins_ldi(t_proc *proc, t_champion *tmp, t_ocp *ocp)
 {
-	unsigned int	p[3];
+	short int		p[3];
 	unsigned int	p_sze[3];
 	unsigned int	ret;
 
@@ -63,10 +63,10 @@ int			cw_ins_ldi(t_proc *proc, t_champion *tmp, t_ocp *ocp)
 	p_sze[2] = cw_ins_param_sze(ocp->third, 2);
 	if (ocp->third != REG_CODE)
 		return (2 + p_sze[0] + p_sze[1] + p_sze[2]);
-	p[0] = cw_ins_ldi_firstparamhandler(proc, tmp, ocp->first);
-	p[1] = cw_ins_ldi_secondparamhandler(proc, tmp, ocp->second, p_sze[0]);
-	p[2] = proc->mem[(tmp->pc + 2 + p_sze[0] + p_sze[1]) % MEM_SIZE];
-	ret = cw_get_data_dir(proc, tmp, (tmp->pc + ((p[0] + p[1]) % IDX_MOD)), 4);
+	p[0] = (short)cw_ins_ldi_firstparamhandler(proc, tmp, ocp->first);
+	p[1] = (short)cw_ins_ldi_secondparamhandler(proc, tmp, ocp->second, p_sze[0]);
+	p[2] = (short)proc->mem[(tmp->pc + 2 + p_sze[0] + p_sze[1]) % MEM_SIZE];
+	ret = (short)cw_get_data_dir(proc, tmp, (tmp->pc + ((p[0] + p[1]) % IDX_MOD)), 4);
 	if (p[2] > 0 && p[2] <= REG_NUMBER)
 	{
 		tmp->reg[p[2] - 1][0] = ret >> 24;
@@ -84,7 +84,7 @@ int			cw_ins_ldi(t_proc *proc, t_champion *tmp, t_ocp *ocp)
 			ft_printf("r%d\n", p[2]);
 		else
 			ft_printf("%d\n", p[2]);
-		ft_printf("%8 | -> load from %d + %d = %d (with pc and mod %u)\n", p[0], p[1], p[0] + p[1], tmp->pc +(p[0] + p[1]) % IDX_MOD);
+		ft_printf("%8 | -> load from %d + %d = %d (with pc and mod %d)\n", p[0], p[1], p[0] + p[1], tmp->pc + (p[0] + p[1]) % IDX_MOD);
 	}
 	return (2 + p_sze[0] + p_sze[1] + p_sze[2]);
 }
